# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Ship Your Playlists API
  version: 1.0.0
  description: |
    API for importing a user's Spotify playlists, finding YouTube matches for
    their tracks, and creating a YouTube playlist from the saved matches.
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
  /api/auth/register:
    post:
      tags: [Authentication]
      summary: Create a user and start a session
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201':
          description: User created. A verification email is sent; session cookies are not set until verification.
          content: { application/json: { schema: { $ref: '#/components/schemas/RegistrationResponse' } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '409': { description: Email is already registered, content: { application/json: { schema: { $ref: '#/components/schemas/Message' } } } }
        '502': { description: Verification email could not be sent, content: { application/json: { schema: { $ref: '#/components/schemas/Message' } } } }
  /api/auth/verify-email:
    get:
      tags: [Authentication]
      summary: Verify a user's email address
      description: The single-use token is sent by email after registration and expires after 24 hours. Successful verification starts a session.
      parameters:
        - { name: token, in: query, required: true, schema: { type: string } }
      responses:
        '200': { description: Email verified and session started, content: { application/json: { schema: { $ref: '#/components/schemas/RefreshResponse' } } } }
        '400': { description: Token is missing, invalid, or expired, content: { application/json: { schema: { $ref: '#/components/schemas/Message' } } } }
  /api/auth/forgot-password:
    post:
      tags: [Authentication]
      summary: Send a password reset link
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ForgotPasswordRequest' }
      responses:
        '200': { description: Generic response that does not reveal whether the email exists, content: { application/json: { schema: { $ref: '#/components/schemas/Message' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
  /api/auth/reset-password:
    post:
      tags: [Authentication]
      summary: Reset a password using an emailed token
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResetPasswordRequest' }
      responses:
        '200': { description: Password reset and session started, content: { application/json: { schema: { $ref: '#/components/schemas/RefreshResponse' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Sign in and start a session
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Signed in. Access and refresh cookies are set.
          headers: { Set-Cookie: { schema: { type: string } } }
          content: { application/json: { schema: { $ref: '#/components/schemas/AuthResponse' } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/auth/get-me:
    get:
      tags: [Authentication]
      summary: Get the current user
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      responses:
        '200': { description: Current user, content: { application/json: { schema: { $ref: '#/components/schemas/CurrentUserResponse' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Deactivated' }
  /api/auth/account:
    delete:
      tags: [Authentication]
      summary: Permanently delete the current user's account
      description: Deletes the account, connected OAuth accounts, cached playlists, tracks, and matches. It does not delete remote Spotify or YouTube playlists. Password accounts must provide the current password.
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeleteAccountRequest' }
      responses:
        '200': { description: Account deleted and session cookies cleared, content: { application/json: { schema: { $ref: '#/components/schemas/Message' } } } }
        '400': { description: Confirmation phrase is missing, content: { application/json: { schema: { $ref: '#/components/schemas/Message' } } } }
        '401': { description: Current password is incorrect, content: { application/json: { schema: { $ref: '#/components/schemas/Message' } } } }
  /api/auth/profile:
    patch:
      tags: [Authentication]
      summary: Update the current user's profile
      description: Update one or more of the user's name, email, and password. Changing the password requires the current password.
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateProfileRequest' }
      responses:
        '200': { description: Profile updated, content: { application/json: { schema: { $ref: '#/components/schemas/CurrentUserResponse' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { description: Current password is incorrect, content: { application/json: { schema: { $ref: '#/components/schemas/Message' } } } }
        '409': { description: Email is already registered, content: { application/json: { schema: { $ref: '#/components/schemas/Message' } } } }
  /api/auth/profile/photo:
    put:
      tags: [Authentication]
      summary: Upload the current user's profile photo
      description: Accepts JPEG, PNG, WebP, or GIF images up to 5 MB in the `photo` multipart field. The saved file is served from the returned `avatarUrl`.
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [photo]
              properties:
                photo:
                  type: string
                  format: binary
      responses:
        '200': { description: Profile photo uploaded, content: { application/json: { schema: { $ref: '#/components/schemas/CurrentUserResponse' } } } }
        '400': { description: Missing, invalid, or oversized image, content: { application/json: { schema: { $ref: '#/components/schemas/Message' } } } }
  /api/auth/refresh-token:
    get:
      tags: [Authentication]
      summary: Refresh session tokens
      description: Requires the `refreshToken` cookie. Replaces both session cookies.
      security: [{ cookieAuth: [] }]
      responses:
        '200': { description: Token refreshed, content: { application/json: { schema: { $ref: '#/components/schemas/RefreshResponse' } } } }
        '401': { description: Refresh token is missing, invalid, or expired, content: { application/json: { schema: { $ref: '#/components/schemas/Message' } } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    cookieAuth: { type: apiKey, in: cookie, name: accessToken }
...
  responses:
    ValidationError: { description: Required input is missing, content: { application/json: { schema: { $ref: '#/components/schemas/Message' } } } }
    Unauthorized: { description: Authentication is missing, invalid, or expired, content: { application/json: { schema: { $ref: '#/components/schemas/Message' } } } }
    Deactivated: { description: Account is deactivated, content: { application/json: { schema: { $ref: '#/components/schemas/Message' } } } }
  schemas:
    Message: { type: object, required: [message], properties: { message: { type: string } } }
    RegisterRequest: { type: object, required: [email, password], properties: { name: { type: string, example: Ada Lovelace }, email: { type: string, format: email, example: ada@example.com }, password: { type: string, format: password, example: correct-horse-battery-staple } } }
    LoginRequest: { type: object, required: [email, password], properties: { email: { type: string, format: email }, password: { type: string, format: password } } }
    DeleteAccountRequest: { type: object, required: [confirmation], properties: { confirmation: { type: string, const: DELETE }, currentPassword: { type: string, format: password } } }
    ForgotPasswordRequest: { type: object, required: [email], properties: { email: { type: string, format: email } } }
    ResetPasswordRequest: { type: object, required: [token, newPassword], properties: { token: { type: string }, newPassword: { type: string, format: password, minLength: 8 } } }
    User: { type: object, required: [id, email, transferredPlaylists], properties: { id: { type: string }, email: { type: string, format: email }, name: { type: [string, 'null'] }, avatarUrl: { type: [string, 'null'], format: uri }, transferredPlaylists: { type: integer, minimum: 0 }, subscription: { type: object }, isActive: { type: boolean } } }
    AuthResponse: { type: object, required: [message, user, accessToken], properties: { message: { type: string }, user: { $ref: '#/components/schemas/User' }, accessToken: { type: string, description: JWT; also returned as an HttpOnly cookie } } }
    RegistrationResponse: { type: object, required: [success, message, user], properties: { success: { type: boolean }, message: { type: string }, user: { $ref: '#/components/schemas/User' }, emailId: { type: [string, 'null'] } } }
    CurrentUserResponse: { type: object, properties: { message: { type: string }, user: { allOf: [{ $ref: '#/components/schemas/User' }, { type: object, properties: { subscription: { type: object }, isActive: { type: boolean } } }] } } }
    UpdateProfileRequest: { type: object, minProperties: 1, properties: { name: { type: string, minLength: 1 }, email: { type: string, format: email }, currentPassword: { type: string, format: password }, newPassword: { type: string, format: password, minLength: 8 } } }
    RefreshResponse: { type: object, properties: { message: { type: string }, accessToken: { type: string } } }
...